Fix: reservations list (reservations.map no iterable error)
- 'ReservationForm' will now have more user feedback when successfully or not
creating a reservation. In addition I am passing the response.payload when
dispatching to 'addReservation' reducer, containing just the reservation data
object.
- 'reservationsSlice' => 'addReservation' reducer and in the
'postReservation.fulfilled' case, will update the 'vehicles' array with
the payload wrap in an array, so the 'vehicles' array will be
iterable in the 'VehiclesList' component.
- 'ReservationsList' will now also check if the 'reservations' is an
empty object and render a message to the user. The reason why the 'reservations'
might be an empty object is because when fetching reservations from the API
and there are no reservations, the API will return an empty object.